Split pkg_manager.py into its own subpackage
