+++ cachito/workers/pkg_managers/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+from cachito.workers.pkg_managers.general import *  # noqa: F401, F403
+from cachito.workers.pkg_managers.golang import *  # noqa: F401, F403
